refactor(client app): redux store, reducers and actions on separate files in a Redux folder,
