+++ client/pages/_app.tsx
@@ -1,10 +1,6 @@
 import '../styles/globals.css';
-import { createStore, applyMiddleware } from 'redux';
+import { store } from '../redux/reducer';
 import { Provider } from 'react-redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import ReduxThunk from 'redux-thunk';
-
-import { state } from '../types/redux';
 
 import Head from 'next/head';
 import { AppProps } from 'next/app';
@@ -13,38 +9,6 @@ import { motion } from 'framer-motion';
 
 import AppLayout from '../layout/Container';
 import { CoordsProvider } from '../hooks/useCoords';
-
-const initialState: state = {
-  isLoading: false,
-  coords: {
-    latitude: 41.395039,
-    longitude: 12.19796
-  },
-  places: []
-};
-function reducer(
-  state = initialState,
-  action: { type: string; payload?: any }
-) {
-  switch (action.type) {
-    case 'SET_COORDINATES':
-      console.log('from coordinates action type:', action.payload);
-      return { ...state, coords: action.payload };
-
-    case 'SET_LOADING':
-      return { ...state, isLoading: action.payload };
-
-    case 'SET_PLACES':
-      return { ...state, places: action.payload };
-
-    default:
-      return state;
-  }
-}
-const store = createStore(
-  reducer,
-  composeWithDevTools(applyMiddleware(ReduxThunk))
-);
 
 function MyApp({ Component, pageProps, router }: AppProps) {
   return (
